+++ sendlib.c
@@ -1174,7 +1174,7 @@ char *mutt_make_date (char *s, size_t len)
   snprintf (s, len,  "Date: %s, %d %s %d %02d:%02d:%02d %+03d%02d\n",
 	    Weekdays[l->tm_wday], l->tm_mday, Months[l->tm_mon],
 	    l->tm_year + 1900, l->tm_hour, l->tm_min, l->tm_sec,
-	    tz / 60, abs (tz) % 60);
+	    (int) tz / 60, (int) abs (tz) % 60);
   return (s);
 }
